Include RetryStat in Runtime
For #191 194
